pyproject.toml
[tool.poetry]
name = "atomhabitsdrfapi"
version = "0.1.0"
description = ""
authors = ["Blagovest Nedkov <purechromas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
celery = "^5.3.4"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
redis = "^5.0.0"
django-celery-beat = "^2.5.0"
django-cors-headers = "^4.2.0"
drf-yasg = "^1.21.7"
pytelegrambotapi = "^4.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
